polish codegen pai
