+++ pkg/sql/codegen/pai/codegen_tf.go
@@ -1,0 +1,142 @@
+// Copyright 2020 The SQLFlow Authors. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package pai
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"strconv"
+	"strings"
+	"text/template"
+
+	"sqlflow.org/sqlflow/pkg/ir"
+	pb "sqlflow.org/sqlflow/pkg/proto"
+	"sqlflow.org/sqlflow/pkg/sql/codegen/tensorflow"
+)
+
+// TFTrainAndSave generates PAI-TF train program.
+func TFTrainAndSave(ir *ir.TrainStmt, session *pb.Session, modelPath string, cc *ClusterConfig) (string, error) {
+	code, err := tensorflow.Train(ir, session)
+	if err != nil {
+		return "", err
+	}
+
+	// append code snippet to save model
+	var tpl = template.Must(template.New("SaveModel").Parse(tfSaveModelTmplText))
+	ckptDir, err := FormatCkptDir(modelPath)
+	if err != nil {
+		return "", err
+	}
+	filler := saveModelFiller{
+		OSSModelDir: ckptDir,
+		Estimator:   ir.Estimator,
+		NumWorkers:  cc.Worker.Count,
+	}
+	var saveCode bytes.Buffer
+	if err = tpl.Execute(&saveCode, filler); err != nil {
+		return "", err
+	}
+	return code + saveCode.String(), nil
+}
+
+// TFLoadAndPredict generates PAI-TF prediction program.
+func TFLoadAndPredict(ir *ir.PredictStmt, session *pb.Session, modelPath string) (string, error) {
+	var tpl = template.Must(template.New("Predict").Parse(tfPredictTmplText))
+	ossModelDir, err := FormatCkptDir(modelPath)
+	if err != nil {
+		return "", err
+	}
+	paiPredictTable := ""
+	if tensorflow.IsPAI() && ir.TmpPredictTable != "" {
+		paiPredictTable = ir.TmpPredictTable
+	}
+	filler := predictFiller{
+		OSSModelDir: ossModelDir,
+		DataSource:  session.DbConnStr,
+		Select:      ir.Select,
+		ResultTable: ir.ResultTable,
+		IsPAI:       tensorflow.IsPAI(),
+		PAITable:    paiPredictTable,
+	}
+	var code bytes.Buffer
+	if err := tpl.Execute(&code, filler); err != nil {
+		return "", err
+	}
+	return code.String(), nil
+}
+
+// TFLoadAndExplain generates PAI-TF explain program.
+func TFLoadAndExplain(ir *ir.ExplainStmt, session *pb.Session, modelPath string) (string, error) {
+	var tpl = template.Must(template.New("Explain").Parse(tfExplainTmplText))
+	ossModelDir, err := FormatCkptDir(modelPath)
+	if err != nil {
+		return "", err
+	}
+	paiExplainTable := ""
+	if tensorflow.IsPAI() && ir.TmpExplainTable != "" {
+		paiExplainTable = ir.TmpExplainTable
+	}
+	filler := explainFiller{
+		OSSModelDir: ossModelDir,
+		DataSource:  session.DbConnStr,
+		Select:      ir.Select,
+		ResultTable: ir.Into,
+		IsPAI:       tensorflow.IsPAI(),
+		PAITable:    paiExplainTable,
+	}
+	var code bytes.Buffer
+	if err := tpl.Execute(&code, filler); err != nil {
+		return "", err
+	}
+	return code.String(), nil
+}
+
+func getTFPAICmd(cc *ClusterConfig, tarball, modelName, ossModelPath, trainTable, valTable, resTable string) (string, error) {
+	jobName := strings.Replace(strings.Join([]string{"sqlflow", modelName}, "_"), ".", "_", 0)
+	cfString, err := json.Marshal(cc)
+	if err != nil {
+		return "", err
+	}
+	cfQuote := strconv.Quote(string(cfString))
+	ckpDir, err := FormatCkptDir(ossModelPath)
+	if err != nil {
+		return "", err
+	}
+
+	// submit table should format as: odps://<project>/tables/<table>,odps://<project>/tables/<table>...
+	submitTables, err := formatODPSTables(trainTable)
+	if err != nil {
+		return "", err
+	}
+	if trainTable != valTable && valTable != "" {
+		valTable, err := formatODPSTables(valTable)
+		if err != nil {
+			return "", err
+		}
+		submitTables = fmt.Sprintf("%s,%s", submitTables, valTable)
+	}
+	outputTables := ""
+	if resTable != "" {
+		table, err := formatODPSTables(resTable)
+		if err != nil {
+			return "", err
+		}
+		outputTables = fmt.Sprintf("-Doutputs=%s", table)
+	}
+	if cc.Worker.Count > 1 {
+		return fmt.Sprintf("pai -name tensorflow1120 -DjobName=%s -Dtags=dnn -Dscript=%s -DentryFile=entry.py -Dtables=%s %s -DcheckpointDir=\"%s\" -Dcluster=%s", jobName, tarball, submitTables, outputTables, ckpDir, cfQuote), nil
+	}
+	return fmt.Sprintf("pai -name tensorflow1120 -DjobName=%s -Dtags=dnn -Dscript=%s -DentryFile=entry.py -Dtables=%s %s -DcheckpointDir=\"%s\"", jobName, tarball, submitTables, outputTables, ckpDir), nil
+}
